@@ -15,32 +15,32 @@
 
 	<properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.release>17</maven.compiler.release>
+        <maven.compiler.release>21</maven.compiler.release>
         <failOnSnapshot>false</failOnSnapshot>
         
         <!-- Manifest information -->
         <maven.build.timestamp.format>yyyy-MM-dd HH:mm</maven.build.timestamp.format>
         <buildDate>${maven.build.timestamp}</buildDate>
         
-		<viewer-core.version>24.06-SNAPSHOT</viewer-core.version>
+		<viewer-core.version>25.01-SNAPSHOT</viewer-core.version>
 		
 		<commons-configuration2.version>2.11.0</commons-configuration2.version>
-        <commons-io.version>2.16.1</commons-io.version>
-        <commons-lang3.version>3.14.0</commons-lang3.version>
-        <eclipselink.version>4.0.3</eclipselink.version>
-		<jakarta.annotation-api.version>1.3.5</jakarta.annotation-api.version>
+        <commons-io.version>2.18.0</commons-io.version>
+        <commons-lang3.version>3.17.0</commons-lang3.version>
+        <eclipselink.version>4.0.5</eclipselink.version>
+		<jakarta.annotation-api.version>3.0.0</jakarta.annotation-api.version>
 		<jakarta.persistence.version>3.2.0</jakarta.persistence.version>
-		<jakarta.servlet-api.version>4.0.4</jakarta.servlet-api.version>
-		<javax.faces.version>2.4.0</javax.faces.version>
-		<json.version>20240303</json.version>
-		<log4j.version>2.23.1</log4j.version>
-        <solr.version>9.6.1</solr.version>
-        <weld-servlet-shaded.version>3.1.9.Final</weld-servlet-shaded.version>
+		<jakarta.servlet-api.version>6.0.0</jakarta.servlet-api.version>
+		<jakarta.faces.version>4.1.2</jakarta.faces.version>
+		<json.version>20250107</json.version>
+		<log4j.version>2.24.3</log4j.version>
+        <solr.version>9.8.0</solr.version>
+        <weld-servlet-shaded.version>4.0.3.Final</weld-servlet-shaded.version>
         
         <!-- Tests -->
         <h2.version>1.4.200</h2.version>
-        <junit.version>5.10.2</junit.version>
-        <mockito.version>5.12.0</mockito.version>
+        <junit.version>5.11.4</junit.version>
+        <mockito.version>5.15.2</mockito.version>
 	</properties>
 	
 	
