@@ -17,8 +17,8 @@ package io.goobi.viewer.managedbeans;
 
 import java.io.Serializable;
 
-import javax.annotation.PostConstruct;
-import javax.inject.Named;
+import jakarta.annotation.PostConstruct;
+import jakarta.inject.Named;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
